Se añadieron validaciones de usuario y correo existente

+++ src/main/java/com/test/blautech/app/controller/CrudController.java
@@ -14,60 +14,64 @@ import com.test.blautech.app.service.IUsuarioService;
 
 @Controller
 public class CrudController {
-	
+
 	@Autowired
-	IUsuarioService usuarioService;
+	private IUsuarioService usuarioService;
 
 	@RequestMapping(value = "/editar", produces = { "application/json" })
 	public @ResponseBody String editarUsuario(@RequestParam(name = "id") Long id,
 			@RequestParam(name = "nickname") String nickname, @RequestParam(name = "email") String email,
 			@RequestParam(name = "password") String password) {
-
-		if (id != null && id > 0) {
-			Usuario usuario = usuarioService.findOne(id);
+		Usuario usuario = new Usuario();
+		if (id != null && id > 0 && usuarioService.findOne(id) != null) {
+			usuario = usuarioService.findOne(id);
 			usuario.setNickname(nickname);
 			usuario.setEmail(email);
 			usuario.setPassword(password);
 			usuarioService.save(usuario);
 			return "Registro actualizado";
 
-		}else {
-			
-		return "No se logro actualizar el registro";
+		} else {
+
+			return "No se logro actualizar el registro";
 		}
-		
-		
+
 	}
+
 	@RequestMapping(value = "/guardar", produces = { "application/json" })
-	public @ResponseBody String guardarUsuario(
-			@RequestParam(name = "nickname") String nickname, @RequestParam(name = "email") String email,
-			@RequestParam(name = "password") String password) {
+	public @ResponseBody String guardarUsuario(@RequestParam(name = "nickname") String nickname,
+			@RequestParam(name = "email") String email, @RequestParam(name = "password") String password) {
 
+		if (usuarioService.findAll().stream().filter(user -> user.getNickname().equals(nickname)).findAny()
+				.orElse(null) == null
+				&& usuarioService.findAll().stream().filter(user -> user.getEmail().equals(email)).findAny()
+						.orElse(null) == null) {
+			Usuario usuario = new Usuario(nickname, password, email);
+			usuarioService.save(usuario);
+			return "Se ha registrado un nuevo usuario";
+		} else {
 
-         Usuario usuario = new Usuario(nickname, password, email);
-         
-         usuarioService.save(usuario);
-         
-         return "Se ha registrado un nuevo usuario";
+			return "el usuario/correo ya existe o no es valido";
+		}
+
 	}
-	
+
 	@RequestMapping(value = "/listar", produces = { "application/json" })
 	public @ResponseBody List<Usuario> listarUsuarios() {
 
 		return usuarioService.findAll();
 
 	}
-	
+
 	@RequestMapping(value = "/eliminar", produces = { "application/json" })
 	public @ResponseBody String eliminar(@RequestParam(name = "id") Long id) {
-		
-		if (id != null && id > 0) {
-			
-			 usuarioService.delete(id);
-			 return "Registro Eliminado";
+
+		if (id != null && id > 0 && usuarioService.findOne(id) != null) {
+
+			usuarioService.delete(id);
+			return "Registro Eliminado";
 		}
 		return "No se ha podido eliminar el registro";
-		
 
 	}
